Refacto: architecture and jwtr
