+++ yaka_scraper/scripts/config.py
@@ -5,4 +5,4 @@ HISTORY_FILE = BASE_PATH / "misc/history"
 META_STORE_FILE_YML = BASE_PATH / "misc/meta_stores.yml"
 META_PROMOTION_FILE_YML = BASE_PATH / "misc/meta_promotion.yml"
 META_BRAND_FILE_YML = BASE_PATH / "misc/meta_brand.yml"
-API_URL = "https://yakosa.herokuapp.com"
+API_URL = "https://yakosa.herokuapp.com//"
