+++ yaka_scraper/scripts/models/store.py
